refactor: `SocketContext` global states -> `recoil` states
