+++ client/src/components/Call/Call.tsx
@@ -1,23 +1,88 @@
-import React from "react";
+import React, { useEffect } from "react";
+import { useRecoilState, useRecoilValue } from "recoil";
 import { Avatar, Typography } from "@mui/joy";
 
 import VideoControls from "../VideoControls/VideoControls";
 import { useSocketContext } from "../../hooks/useSocketContext";
 import useStyles from "./styles";
+import {
+  callAtom,
+  callEndedAtom,
+  isCallerMutedAtom,
+  isVideoAtom,
+  joinedRoomAtom,
+  nameAtom,
+  streamAtom,
+} from "../../atoms";
+import { peer, socket } from "../../utils";
 
 const Call = (): JSX.Element => {
-  const {
-    isVideo,
-    callEnded,
-    myVideo,
-    userVideo,
-    stream,
-    joinedRoom,
-    call,
-    isCallerMuted,
-  } = useSocketContext();
+  const { myVideo, userVideo } = useSocketContext();
+
+  const [call, setCall] = useRecoilState(callAtom);
+
+  const callEnded = useRecoilValue(callEndedAtom);
+  const isCallerMuted = useRecoilValue(isCallerMutedAtom);
+  const isVideo = useRecoilValue(isVideoAtom);
+  const stream = useRecoilValue(streamAtom);
+  const joinedRoom = useRecoilValue(joinedRoomAtom);
+  const name = useRecoilValue(nameAtom);
 
   const classes = useStyles();
+
+  useEffect(() => {
+    peer.on("connection", ({ metadata }) => {
+      if ("name" in metadata) {
+        setCall((prevState) => ({
+          ...prevState,
+          name: metadata.name,
+        }));
+      } else if ("isVideo" in metadata) {
+        setCall((prevState) => ({
+          ...prevState,
+          isVideo: metadata.isVideo,
+        }));
+      }
+    });
+
+    socket.on("user-disconnected", () => {
+      setCall({
+        isRecievedCall: false,
+        from: "",
+        name: "",
+        isVideo: true,
+      });
+      userVideo.current!.srcObject = null;
+    });
+  }, []);
+
+  useEffect(() => {
+    peer.on("call", (incomingCall) => {
+      setCall((prevState) => ({
+        ...prevState,
+        from: incomingCall.peer,
+        isRecievedCall: true,
+        name: incomingCall.metadata.name,
+      }));
+
+      incomingCall.answer(stream);
+      myVideo.current!.srcObject = stream!;
+
+      incomingCall.on("stream", (currentStream) => {
+        userVideo.current!.srcObject = currentStream;
+        peer.connect(incomingCall.peer, { metadata: { name } });
+      });
+    });
+  }, [name]);
+
+  useEffect(() => {
+    if (myVideo.current) myVideo.current!.srcObject = stream!;
+    peer.connect(call.from, { metadata: { isVideo } });
+  }, [isVideo]);
+
+  useEffect(() => {
+    if (joinedRoom && myVideo.current) myVideo.current!.srcObject = stream!;
+  }, [joinedRoom]);
 
   return (
     <div className={classes.videoContainer}>
